fix: custom component cart
