+++ src/pages/customer/cart-page/components/__mocks__/data.ts
@@ -1,0 +1,33 @@
+import shirt_1 from '@/assets/images/shirt-1.png';
+import shirt_2 from '@/assets/images/shirt-2.png';
+import shirt_3 from '@/assets/images/shirt-3.png';
+
+export const LIST_CART = [
+  {
+    id: 'xxxx',
+    amount: 1,
+    price: 145,
+    size: 'Small',
+    color: 'White',
+    name: 'Gradient Graphic T-shirt',
+    src: shirt_1,
+  },
+  {
+    id: 'yyyy',
+    amount: 1,
+    price: 180,
+    size: 'Mall',
+    color: 'White',
+    name: 'Checkered Shirt',
+    src: shirt_2,
+  },
+  {
+    id: 'zzzz',
+    amount: 1,
+    price: 240,
+    size: 'Large',
+    color: 'White',
+    name: 'Skinny Fit Jeans',
+    src: shirt_3,
+  },
+];
