+++ src/pages/customer/cart-page/components/index.ts
@@ -1,0 +1,2 @@
+export * from './ListProduct';
+export * from './Payment';
