+++ src/pages/customer/index.ts
@@ -1,2 +1,3 @@
 export * from './home-page';
 export * from './product-detail';
+export * from './cart-page';
